requirements.txt
pyspark==3.5.0
delta-spark==3.0.0
boto3==1.34.0
psycopg2-binary==2.9.9
google-api-python-client==2.110.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-auth==2.23.4
pandas==2.1.4
pytz==2023.3
python-dotenv==1.0.0
structlog==23.2.0
psutil==5.9.6